Try bundling depecencies with deno esbuild + tweak to esbuild http import plugin
This a first step to try to handle peer dependecies in order to use the UI library alongside the consumer's imported react version.
Wich require import both the library and react with the same cdn.
In order to try this bundle, this commits create a new version of the package will be then imported with esm in a second time to test if imports are working